clean the preprocess process to fit in one file - readability is low but to be fixed in functions at least
