--- data/preprocess/regex_desc/handle_salary.py
@@ -1,5 +1,0 @@
-import pandas as pd 
-import re
-import os
-
-data = pd.read_csv('..\Scrapping\data\complete\indeed_jobs.csv')
